% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miete.R
\docType{data}
\name{miete}
\alias{miete}
\title{miete dataset}
\format{
A data frame with 2053 observations on the following 12 variables.
\describe{
  \item{rent}{Rent in euros.}
  \item{bathextra}{Special furniture in bathroom, yes = 1, no = 0.}
  \item{tiles}{Bathroom with tiles, yes = 0, no = 1.}
  \item{area}{Municipality.}
  \item{kitchen}{Upmarket kitchen, yes = 1, no = 0.}
  \item{rooms}{Number of rooms.}
  \item{best}{Best address, yes = 1, no = 0.}
  \item{good}{Good address, yes = 1, no =0.}
  \item{warm}{Warm water, yes = 0, no = 1.}
  \item{central}{Central heating, yes = 0, no = 1.}
  \item{year}{Year of construction.}
  \item{size}{Living space in square meter.}
}
}
\usage{
data(miete)
}
\description{
The miete data contains the rent index for Munich in 2003.
}
\examples{
data(miete)
summary(miete)
}
\references{
Fahrmeir, L., Künstler, R., Pigeot, I., Tutz, G. (2004) Statistik: der Weg zur Datenanalyse. 5. Auflage, Berlin: Springer-Verlag.
}
\keyword{datasets}
