% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.DMR.R
\name{plot.DMR}
\alias{plot.DMR}
\title{plot.DMR}
\usage{
\method{plot}{DMR}(x, ...)
}
\arguments{
\item{x}{Fitted DMR object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plot coefficients from a DMR object.
}
\details{
Produces a coefficient profile plot of the coefficient paths for a fitted DMR object.
}
\examples{
data(miete)
y <- miete[,1]
X <- miete[,-1]
m <- DMR(X, y)
plot(m)
}

