% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.DMR.R
\name{print.DMR}
\alias{print.DMR}
\title{print.DMR}
\usage{
\method{print}{DMR}(x, ...)
}
\arguments{
\item{x}{Fitted DMR object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The summary is silently returned.
}
\description{
Print a DMR object.
}
\details{
Print a summary of the DMR path at each step along the path.
}
\examples{
data(miete)
y <- miete[,1]
X <- miete[,-1]
m <- DMR(X, y)
print(m)
}

