% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv.DMR.R
\name{plot.cv.DMR}
\alias{plot.cv.DMR}
\title{plot.cv.DMR}
\usage{
\method{plot}{cv.DMR}(x, ...)
}
\arguments{
\item{x}{Fitted cv.DMR object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plot cross-validated error values from a cv.DMR object.
}
\details{
Produces a plot of cross-validated error values for the entire sequence of models from the fitted cv.DMR object.
}
\examples{
## cv.DMR for linear regression
set.seed(13)
data(miete)
y <- miete$rent
X <- miete$area
cv = cv.DMR(X,y)
plot(cv)

}

