% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gic.DMR.R
\name{plot.gic.DMR}
\alias{plot.gic.DMR}
\title{plot.gic.DMR}
\usage{
\method{plot}{gic.DMR}(x, ...)
}
\arguments{
\item{x}{Fitted gic.DMR object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plot gic values from a gic.DMR object.
}
\details{
Produces a plot of generalized information criterion for the entire sequence of models from the fitted gic.DMR object.
}
\examples{
data(miete)
y <- miete[,1]
X <- miete[,-1]
m <- DMR(X, y)
g <- gic.DMR(m, c = 2.5)
plot(g)
}

